chore: move proxy utils to src/server/fetcher/
